Move some unit tests.
Some of the unit tests in `polys/numberfields/tests/test_numbers` really belonged in `~/test_minpoly`.
